+++ src/lib/components/common/InfiniteScrollTrigger.tsx
@@ -1,4 +1,5 @@
-import { useCallback, useEffect, useRef } from "react"
+import { useRef } from "react"
+import { useIntersectionObserver } from "$lib/hooks"
 import LoadingSpinner from "./LoadingSpinner"
 import "./style/infiniteScrollTrigger.scss"
 
@@ -9,41 +10,13 @@ type LoadMoreProps = {
 
 export default function InfiniteScrollTrigger(props: LoadMoreProps) {
     const { handleLoadMore, hasNextPage } = props
+
+    const onIntersect = () => {
+       hasNextPage && handleLoadMore()
+    }
+
 	const sentinelRef = useRef<HTMLDivElement>(null);
-	const observerRef = useRef<IntersectionObserver | null>(null);
-    
-
-    const handleIntersect = useCallback((entries: IntersectionObserverEntry[]) => {
-        if (entries[0].isIntersecting && hasNextPage) {
-            handleLoadMore()
-        }
-    }, [hasNextPage, handleLoadMore])
-
-	useEffect(() => {
-		observerRef.current = new IntersectionObserver(handleIntersect, {
-			root: null,
-			rootMargin: "0px",
-			threshold: 0,
-		});
-
-		if (sentinelRef.current) {
-			observerRef.current.observe(sentinelRef.current)
-		}
-
-		return () => {
-			if (observerRef.current) {
-				observerRef.current.disconnect()
-			}
-		}
-	}, [handleIntersect]);
-
-    useEffect(() => {
-		if (observerRef.current && sentinelRef.current) {
-			observerRef.current.disconnect();
-			observerRef.current.observe(sentinelRef.current);
-		}
-	}, [hasNextPage]);
-
+    useIntersectionObserver(sentinelRef, onIntersect)
 
     return (
         <div className="infinite-scroll-trigger-component">
